% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chimpanzee.r
\docType{data}
\name{chimpanzee}
\alias{chimpanzee}
\title{Pan troglodytes (chimpanzee) gene transfer format file data}
\format{A data frame with 529537 rows and 9 variables:
\describe{
\url{http://www.gencodegenes.org/gencodeformat.html}
}}
\usage{
data(chimpanzee)
}
\description{
A dataset containing annotations of both coding and non-coding genes
of Pan_troglodytes.CHIMP2.1.4.84.gtf
}
\keyword{datasets}

