% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cow.r
\docType{data}
\name{cow}
\alias{cow}
\title{Bos taurus (cow) gene transfer format file data}
\format{A data frame with 563844 rows and 9 variables:
\describe{
\url{http://www.gencodegenes.org/gencodeformat.html}
}}
\usage{
data(cow)
}
\description{
A dataset containing annotations of both coding and non-coding genes
of Bos_taurus.UMD3.1.84.gtf
}
\keyword{datasets}

