% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fly.r
\docType{data}
\name{fly}
\alias{fly}
\title{Drosophila melanogaster (fly) gene transfer format file data}
\format{A data frame with 538684 rows and 9 variables:
\describe{
\url{http://www.gencodegenes.org/gencodeformat.html}
}}
\usage{
data(fly)
}
\description{
A dataset containing annotations of both coding and non-coding genes
of Drosophila_melanogaster.BDGP6.84.gtf
}
\keyword{datasets}

