% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/frog.r
\docType{data}
\name{frog}
\alias{frog}
\title{Xenopus tropicalis (frog) gene transfer format file data}
\format{A data frame with 581787 rows and 9 variables:
\describe{
\url{http://www.gencodegenes.org/gencodeformat.html}
}}
\usage{
data(frog)
}
\description{
A dataset containing annotations of both coding and non-coding genes
of Xenopus_tropicalis.JGI_4.2.84.gtf
}
\keyword{datasets}

