% Generated by roxygen2 (4.0.1): do not edit by hand
\docType{class}
\name{GTF-class}
\alias{GTF}
\alias{GTF-class}
\title{Reference class to manipulate GTF data}
\description{
Reference class to manipulate GTF data
}
\section{Fields}{

\describe{
\item{\code{GTF}}{complicated structure of list}

\item{\code{sorted}}{data is sorted or not}
}}
\section{Methods}{

\describe{
\item{\code{geneLength(type = c("gene", "exon"))}}{get gene/exon length}

\item{\code{getGeneID(index = NULL)}}{get Gene ID (ensembl ID)}

\item{\code{getTranscriptsByGeneID(geneID)}}{get transcripts by gene IDs}

\item{\code{getValueByGeneID(geneID, type = c("chr", "name", "strand", "type", "end",
  "start"))}}{get values corresponding to gene IDs}

\item{\code{mergeTranscripts(mc.cores = 1)}}{merge alternative transcripts into one union gene}

\item{\code{read(gtf)}}{read from GTF file}

\item{\code{sort()}}{sort GTF}

\item{\code{toBed(file = NULL, type = c("gene", "exon"), chromosome = paste("chr",
  c(1:22, "X", "Y"), sep = ""))}}{convert and write to BED file}

\item{\code{write(file)}}{write to GTF file}
}}
\examples{
\dontrun{
library(GTF)
gencode = new("GTF")
gencode$read(paste(system.file(package = "GTF"), "/extdata/test.gtf", sep=""))
gencode$mergeTranscripts()
gencode$mergeTranscripts(mc.cores = 2)
gencode$write("output.gtf")
gene_length = gencode$geneLength("exon")
gene = gencode$toBed(type = "gene")
exon = gencode$toBed(type = "exon")
gencode$toBed(file = "gene.bed", type = "gene")
gencode$toBed(file = "exon.bed", type = "exon")
gi = gencode$getGeneID(1:10)
chr = gencode$getValueByGeneID(gi, type = "chr")
tr = gencode$getTranscriptsByGeneID(gi)
}
}

