% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/human.r
\docType{data}
\name{human}
\alias{human}
\title{Homo sapiens (human) gene transfer format file data}
\format{A data frame with 2619444 rows and 9 variables:
\describe{
\url{http://www.gencodegenes.org/gencodeformat.html}
}}
\usage{
data(human)
}
\description{
A dataset containing annotations of both coding and non-coding genes
of gencode.v19.annotation.gtf
}
\keyword{datasets}

