% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mouse.r
\docType{data}
\name{mouse}
\alias{mouse}
\title{Mus musculus (mouse) gene transfer format file data}
\format{A data frame with 1589642 rows and 9 variables:
\describe{
\url{http://www.gencodegenes.org/gencodeformat.html}
}}
\usage{
data(mouse)
}
\description{
A dataset containing annotations of both coding and non-coding genes
of Mus_musculus.GRCm38.84.gtf
}
\keyword{datasets}

