% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pig.r
\docType{data}
\name{pig}
\alias{pig}
\title{Sus scrofa (pig) gene transfer format file data}
\format{A data frame with 591710 rows and 9 variables:
\describe{
\url{http://www.gencodegenes.org/gencodeformat.html}
}}
\usage{
data(pig)
}
\description{
A dataset containing annotations of both coding and non-coding genes
of Sus_scrofa.Sscrofa10.2.84.gtf
}
\keyword{datasets}

