% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rat.r
\docType{data}
\name{rat}
\alias{rat}
\title{Rattus norvegicus (rat) gene transfer format file data}
\format{A data frame with 746468 rows and 9 variables:
\describe{
\url{http://www.gencodegenes.org/gencodeformat.html}
}}
\usage{
data(rat)
}
\description{
A dataset containing annotations of both coding and non-coding genes
of Rattus_norvegicus.Rnor_6.0.82.gtf
}
\keyword{datasets}

