% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/worm.r
\docType{data}
\name{worm}
\alias{worm}
\title{Caenorhabditis elegans (worm) gene transfer format file data}
\format{A data frame with 662705 rows and 9 variables:
\describe{
\url{http://www.gencodegenes.org/gencodeformat.html}
}}
\usage{
data(worm)
}
\description{
A dataset containing annotations of both coding and non-coding genes
of Caenorhabditis_elegans.WBcel235.84.gtf
}
\keyword{datasets}

