% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/yeast.r
\docType{data}
\name{yeast}
\alias{yeast}
\title{Saccharomyces cerevisiae (yeast) gene transfer format file data}
\format{A data frame with 42071 rows and 9 variables:
\describe{
\url{http://www.gencodegenes.org/gencodeformat.html}
}}
\usage{
data(yeast)
}
\description{
A dataset containing annotations of both coding and non-coding genes
of Saccharomyces_cerevisiae.R64-1-1.84.gtf
}
\keyword{datasets}

