% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zebrafish.r
\docType{data}
\name{zebrafish}
\alias{zebrafish}
\title{Danio rerio (zebrafish) gene transfer format file data}
\format{A data frame with 1142757 rows and 9 variables:
\describe{
\url{http://www.gencodegenes.org/gencodeformat.html}
}}
\usage{
data(zebrafish)
}
\description{
A dataset containing annotations of both coding and non-coding genes
of Danio_rerio.GRCz10.84.gtf
}
\keyword{datasets}

